refactor: movies page module
